feat: version sync (#16)
Adds a script to check version in CHANGELOG.md against version in __version__.py or overwrite the latter version. There's an optional variable to specify the location of __version__.py. Make targets are added, and the check is added to the make check target, and consequently becomes part of CI.

Co-authored-by: cragwolfe <crag@unstructuredai.io>