fix: standard import ref's accessible in api cells (#156)
Before this fix, one could not import os in a # pipeline-api as the resulting FastAPI python module would then have import os defined twice. But, in the case where# pipeline-api cell had a reference to an os module that had been imported previously in a non-# pipeline-api cell, an error occurred when converting the notebook since os had not been defined (as only # pipeline-api cells are inspected, per design).

The fix is to include implied "standard" imports for os, gzip, mimetypes, and json when inferring parameters in pipeline_api, but not when executing the jinja pipeline-template logic since those imports already exist at the top of the template.