Handle case where file.content_type is None
I can't reproduce this with the Starlette test client, but the following snippet in a python shell
is returning a 400 error from the hosted api.

```
import requests

filename="fake.doc"
with open(filename, "rb") as f:
    response = requests.post(
        "https://api.unstructured.io/general/v0/general",
        files={"files": (str(filename), f)},
    )

print(response.text)
'''

When we don't specify the content type in requests, it's causing the FastAPI file.content_type to be
None, at least on my machine.