Filter out logs on /healthcheck
Reduce noisy uvicorn logs when polling /healthcheck.

To test, you can run a local copy of api-tools to regenerate one of the pipeline apis, and try
calling /healthcheck on it.

For instance, in api-tools, run `make install-project-local`. Jump over to unstructured-api and run
`make generate-api` and `make run-web-app`. Make a request to localhost:8000/healthcheck and verify
that the uvicorn logs do not show it.